Start of binary serialization implementation: - added BinarySerializable base class for serializable objects - added binaryostream to handle serialization - added endianness helpers
Added custom exception class:
- ArgumentException

Fixed compilation in asio 1.11.0

Very little work on server and session (unbinding server, moved binding from constructor to bind method, run function, StopAcceptor function, session id is now uint64_t)

TODO:
- w pizdu rzeczy

+++ include/session.hpp
@@ -23,7 +23,7 @@ namespace snesl
 
         asio::ip::tcp::socket & GetSocket();
 
-        void ConnectionStarted(unsigned long long id);
+        void ConnectionStarted(uint64_t id);
 
         void HandleReadSize(const asio::error_code& error, const size_t& /*bytes*/);
         void HandleReadMsg(const asio::error_code& error, const size_t& /*bytes*/);
